refactor: AuthenticationTokenConverter 에 인증정보 추출 및 변환 역할을 담당하도록 변경
