+++ src/test/java/nextstep/security/authentication/token/BasicAuthenticationTokenConverterTest.java
@@ -1,27 +1,28 @@
-package nextstep.security.authentication.token.basic;
+package nextstep.security.authentication.token;
 
+import jakarta.servlet.http.HttpServletRequest;
 import nextstep.security.authentication.Authentication;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Base64;
+import org.springframework.mock.web.MockHttpServletRequest;
 
 import static nextstep.security.Fixture.PASSWORD;
 import static nextstep.security.Fixture.USERNAME;
+import static nextstep.security.Fixture.createBasicRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-class BasicAuthenticationTokenTest {
+class BasicAuthenticationTokenConverterTest {
+
+    private final AuthenticationTokenConverter converter = BasicAuthenticationTokenConverter.getInstance();
 
     @DisplayName("Basic token 을 디코딩할 수 있다.")
     @Test
     void of() {
-        final String basicToken = encode(USERNAME, PASSWORD);
-        final Authentication authentication = BasicAuthenticationToken.of(
-                encode(USERNAME, PASSWORD)
-        );
+        final HttpServletRequest request = createBasicRequest();
+        final Authentication authentication = converter.convert(request);
         assertAll(
-                () -> assertThat(BasicAuthenticationToken.supports(basicToken))
+                () -> assertThat(converter.supports(request))
                         .isTrue(),
                 () -> assertThat(authentication.getPrincipal())
                         .isEqualTo(USERNAME),
@@ -33,13 +34,10 @@ class BasicAuthenticationTokenTest {
     @DisplayName("잘못된 Basic token 은 지원하지 않는다.")
     @Test
     void notSupported() {
-        assertThat(BasicAuthenticationToken.supports("invalid"))
+        final MockHttpServletRequest request = new MockHttpServletRequest();
+        request.addHeader("Authorization", "invalid token");
+        request.setRequestURI("/members");
+        assertThat(converter.supports(request))
                 .isFalse();
-    }
-
-    private String encode(String username, String password) {
-        return "Basic " + Base64.getEncoder().encodeToString(
-                (username + ":" + password).getBytes()
-        );
     }
 }
